Revising Homepage
1. Hover Effect on Button:
Added a hover effect on the button to change its background color and text color.

2. Change Lorem Ipsum Text:
Replace the placeholder text with a more meaningful description.

3. Change Carousel Photos:
Replace the images in the carousel with new ones.

4. Change Slider Partners Photos:
Update the images in the partners' slider at the bottom.

5. Shorten Button Text in Cards:
Make the button text in cards shorter.

6. Add Ellipsis for Long Descriptions and Titles in Cards:
If the description is more than 15 words or the title is more than 10 words in cards, add an ellipsis (...).